separate coverage logic
